src/core/report_generator.py
"""
Generate PDF and Word reports for scan results
"""
from datetime import datetime
from reportlab.lib import colors
from reportlab.lib.pagesizes import letter
from reportlab.platypus import SimpleDocTemplate, Table, TableStyle
from docx import Document

class ReportGenerator:
    def __init__(self):
        self.timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
    
    def generate_pdf(self, scan_results, risk_analysis, output_path=None):
        """Generate PDF report"""
        if output_path is None:
            output_path = f"scan_report_{self.timestamp}.pdf"
            
        doc = SimpleDocTemplate(output_path, pagesize=letter)
        elements = []
        
        # Add scan results table
        # Implementation details here
        
        doc.build(elements)
        return output_path
    
    def generate_word(self, scan_results, risk_analysis, output_path=None):
        """Generate Word report"""
        if output_path is None:
            output_path = f"scan_report_{self.timestamp}.docx"
            
        doc = Document()
        doc.add_heading('Port Scan Report', 0)
        
        # Add scan results
        # Implementation details here
        
        doc.save(output_path)
        return output_path