requirements.txt
# Core dependencies
PyQt6==6.6.1
nmap-python==0.7.1
reportlab==4.1.0
python-docx==1.1.0
faiss-cpu==1.7.4
numpy==1.26.3
requests==2.31.0
mistralai==0.0.12

# Development dependencies
black==24.1.1
pylint==3.0.3
pytest==8.0.0